+++ src/MagicOnion.Server/Binder/IMagicOnionGrpcMethod.cs
@@ -1,14 +1,19 @@
 using System.Reflection;
+using Grpc.Core;
 
 namespace MagicOnion.Server.Binder;
 
-public interface IMagicOnionGrpcMethod;
+public interface IMagicOnionGrpcMethod
+{
+    MethodType MethodType { get; }
+    Type ServiceType { get; }
+    string ServiceName { get; }
+    string MethodName { get; }
+    MethodInfo MethodInfo { get; }
+}
 
 public interface IMagicOnionGrpcMethod<TService> : IMagicOnionGrpcMethod
     where TService : class
 {
-    string ServiceName { get; }
-    string MethodName { get; }
-    MethodInfo MethodInfo { get; }
     void Bind(IMagicOnionGrpcMethodBinder<TService> binder);
 }
