--- src/MagicOnion.Server/Binder/Legacy/MagicOnionService.cs
@@ -1,4 +1,0 @@
-namespace MagicOnion.Server.Binder;
-
-// The MagicOnion service methods are bound by `MagicOnionServiceMethodProvider<TService>`
-internal class MagicOnionService;
