--- src/MagicOnion.Server/Binder/Legacy/MagicOnionServiceMethodProvider.cs
@@ -1,24 +1,0 @@
-using Grpc.AspNetCore.Server.Model;
-
-namespace MagicOnion.Server.Binder;
-
-[Obsolete]
-internal class MagicOnionServiceMethodProvider<TService> : IServiceMethodProvider<TService>
-    where TService : class
-{
-    readonly MagicOnionServiceDefinition magicOnionServiceDefinition;
-
-    public MagicOnionServiceMethodProvider(MagicOnionServiceDefinition magicOnionServerServiceDefinition)
-    {
-        magicOnionServiceDefinition = magicOnionServerServiceDefinition ?? throw new ArgumentNullException(nameof(magicOnionServerServiceDefinition));
-    }
-
-    public void OnServiceMethodDiscovery(ServiceMethodProviderContext<TService> context)
-    {
-        var binder = new MagicOnionServiceBinder<TService>(context);
-        foreach (var methodHandler in magicOnionServiceDefinition.MethodHandlers)
-        {
-            methodHandler.BindHandler(binder);
-        }
-    }
-}
