+++ src/MagicOnion.Server/Extensions/MagicOnionEndpointRouteBuilderExtensions.cs
@@ -6,14 +6,12 @@ namespace Microsoft.AspNetCore.Builder;
 
 public static class MagicOnionEndpointRouteBuilderExtensions
 {
-    public static GrpcServiceEndpointConventionBuilder MapMagicOnionService(this IEndpointRouteBuilder builder)
+    public static void MapMagicOnionService(this IEndpointRouteBuilder builder)
     {
         var context = new MagicOnionGrpcServiceRegistrationContext(builder);
         foreach (var methodProvider in builder.ServiceProvider.GetRequiredService<IEnumerable<IMagicOnionGrpcMethodProvider>>())
         {
             methodProvider.OnRegisterGrpcServices(context);
         }
-
-        return builder.MapGrpcService<MagicOnionService>();
     }
 }
