+++ src/MagicOnion.Server/ServiceContext.Streaming.cs
@@ -1,6 +1,7 @@
 using System.Reflection;
 using Grpc.Core;
 using MagicOnion.Serialization;
+using MagicOnion.Server.Binder;
 using MagicOnion.Server.Diagnostics;
 using MagicOnion.Server.Internal;
 using MessagePack;
@@ -40,25 +41,21 @@ internal class StreamingServiceContext<TRequest, TResponse> : ServiceContext, IS
 
     public StreamingServiceContext(
         object instance,
-        Type serviceType,
-        string serviceName,
-        MethodInfo methodInfo,
+        IMagicOnionGrpcMethod method,
         ILookup<Type, Attribute> attributeLookup,
-        MethodType methodType,
         ServerCallContext context,
         IMagicOnionSerializer messageSerializer,
         ILogger logger,
-        MethodHandler methodHandler,
         IServiceProvider serviceProvider,
         IAsyncStreamReader<TRequest>? requestStream,
         IServerStreamWriter<TResponse>? responseStream
-    ) : base(instance, serviceType, serviceName, methodInfo, attributeLookup, methodType, context, messageSerializer, logger, methodHandler, serviceProvider)
+    ) : base(instance, method, attributeLookup, context, messageSerializer, logger, serviceProvider)
     {
         RequestStream = requestStream;
         ResponseStream = responseStream;
 
         // streaming hub
-        if (methodType == MethodType.DuplexStreaming)
+        if (MethodType == MethodType.DuplexStreaming)
         {
             this.streamingResponseWriter = new Lazy<QueuedResponseWriter<TResponse>>(() => new QueuedResponseWriter<TResponse>(this));
         }
